move PropertyKey to root package
it's not specific to traversal, just happens to be a scala case class,
and there's currently no better subproject
